% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_branch_features}
\alias{compute_branch_features}
\title{Compute branch features}
\usage{
compute_branch_features(ifile, omitapical = FALSE, omitaxon = FALSE,
  omitdend = FALSE, correct = FALSE, selection = "all")
}
\description{
Compute branch features
}
