% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_node_features}
\alias{compute_node_features}
\title{Compute node features}
\usage{
compute_node_features(ifile, omitapical = FALSE, omitaxon = FALSE,
  omitdend = FALSE, correct = FALSE)
}
\description{
Compute node features
}
