% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert}
\alias{convert}
\title{Converts a reconstruction to SWC or JSON.}
\usage{
convert(ifile, ext, eps = 0, correct = FALSE)
}
\description{
Use resolved paths, i.e., not with ~ for the home directory.
}
