% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{json2dataframe}
\alias{json2dataframe}
\title{JSON to dataframe.}
\usage{
json2dataframe(json)
}
\description{
For feature extraction, adds all the measures as dataframe columns.
For validation, it omits any detailed results.
}
