% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{merge_branch_node}
\alias{merge_branch_node}
\title{Merge branch and node features into a single data frame.
introduces dependency on dplyr.}
\usage{
merge_branch_node(b, n)
}
\description{
Merge branch and node features into a single data frame.
introduces dependency on dplyr.
}
